--- numba_dpex/experimental/kernel_iface/kernel_sim/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2023 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
